% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/back_up_gh.R
\name{back_up_gh_orgs}
\alias{back_up_gh_orgs}
\title{Back up GitHub organization repositories}
\usage{
back_up_gh_orgs(gh_token, backup_path, orgs_name = NULL)
}
\arguments{
\item{gh_token}{Token number from GitHub website.}

\item{backup_path}{Directory where backup materials are located.}

\item{orgs_name}{Name of the organization. A vector of names. Otherwise
the function will backup all organizations that are associated with
your account.}
}
\value{
Backup of GitHub repository in the directory where \code{backup_path} is
}
\description{
The function uses gh_token to access GitHub API URL to download organization
repositories. It returns all backup materials in a folder on users'
local machine.
}
\examples{
\dontrun{
back_up_gh_orgs(
  gh_token = "abcdefghijklmnopqrstuvwxyz01234567890123",
  backup_path = "C:/Users/ghbackup/",
  orgs_name = c("nmfs-fish-tools", "nmfs-general-modeling-tools")
)

back_up_gh_orgs(
  gh_token = gh::gh_token(),
  backup_path = "C:/Users/ghbackup/",
  orgs_name = c("nmfs-fish-tools", "nmfs-general-modeling-tools")
)
}
}
