% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.R
\name{sum}
\alias{sum}
\title{Sum of two or more numbers.}
\usage{
sum(x)
}
\arguments{
\item{x}{A vector of numbers. Minimum requirements: two numbers.}
}
\value{
The sum of vector \code{x}.
}
\description{
Add a vector of numbers and return the sum of the vector of numbers.
}
\examples{
sum(c(1, 2))
sum(c(4, 5, 6, 7))
}
